desktop_app/requirements.txt
pywebview>=4.0
requests>=2.28.0
pyinstaller>=5.0
